# paddlelabel/openapi.yml
openapi: 3.0.0
info:
  title: PaddleLabel API Specs
  version: 1.0.2
  description: Web backend APIs for PaddleLabel
  contact:
    name: PaddleLabel Developers
    url: 'https://github.com/PaddleCV-SIG/PaddleLabel/issues'
    email: me@linhan.email
  license:
    name: Apache License 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
servers:
  - url: /api
    description: Same origin
paths:
...
  '/projects/{project_id}/labels':
    parameters:
      - $ref: '#/components/parameters/project_id'
    get:
      summary: Get all labels under a project
      tags:
        - Project
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
      operationId: getLabels
      description: ''
    post:
      summary: 'Set all labels under a project, will delete previous labels'
      tags:
        - Project
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
      operationId: setLabels
      description: ''
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Label'
    delete:
      summary: Delete all labels under a project
      tags:
        - Project
      responses:
        '200':
          description: success
      operationId: removeLabels
      description: ''
...
  /labels:
    get:
      tags:
        - Label
      summary: 'Get all labels, sort by last modify'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
      operationId: getAll
    post:
      tags:
        - Label
      summary: Create new labels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Label'
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Label'
      parameters:
        - $ref: '#/components/parameters/request_id'
        - schema:
            type: string # TODO: 原来是boolean，现在是string，解析后无法获取到此字段
          in: header
          name: remove_duplicate_by_name
          description: 是否根据标签名去重，标签名如果存在不会创建
      operationId: create
  '/labels/{label_id}':
    parameters:
      - $ref: '#/components/parameters/label_id'
    get:
      tags:
        - Label
      summary: Get info about a specific label
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        '404':
          description: Label not fond
      operationId: get
    delete:
      summary: Delete a label
      description: Delete a label. Labels in use (have annotation pointing at them) are not allowed to be deleted.
      tags:
        - Label
      responses:
        '200':
          description: OK
        '404':
          description: The label specified is not found
      operationId: remove
    put:
      tags:
        - Label
      summary: Edit label info
      description: Edit label info. Provide key value pair to change one value only. Provide all changed values to change multiple. Empty string will be set. Leave values don't intend to change out of request body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Label'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        '404':
          description: 'Label with specified label id not fond, or project dont have requested property'
      operationId: update
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: paddlelabel.api.decode_token
  schemas:
...
    Label:
      title: Label
      type: object
      description: ''
      properties:
        label_id:
          type: integer
        project_id:
          type: integer
        name:
          type: string
        color:
          type: string
        comment:
          type: string
        created:
          type: string
          readOnly: true
        modified:
          type: string
          readOnly: true
        super_category_id:
          type: integer
      required:
        - name
...
  parameters:
    request_id:
      name: request_id
      in: header
      required: false
      schema:
        type: integer
        maxLength: 30
      description: 'Assign a unique random string each request, backend will reject requests with same request_id within several seconds. Prevent critical operations (mostly post) from executing multiple times'
    project_id:
      name: project_id
      in: path
      schema:
        type: integer
      description: Id of a project
      required: true
    label_id:
      name: label_id
      in: path
      required: true
      schema:
        type: integer
      description: Id of a label
...